+++ src/load/load_raw_map.c
@@ -9,7 +9,7 @@
 #include "libft/t_list.h"
 #include "log.h"
 #include "t_map.h"
-#include "t_map_internals.h"
+#include "load_internals.h"
 
 static t_list* load_lines_in_lst(const char* map_path);
 static t_map move_str_list_to_map(t_list** str_lst_ref);
