+++ src/load/parse_map.c
