+++ src/load/t_charset.c
